% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lavaan_analysis_tables.R
\name{make_lavaan_analysis_tables}
\alias{make_lavaan_analysis_tables}
\title{Comprehensive Lavaan Regression Analysis}
\usage{
make_lavaan_analysis_tables(data, dvs, ivs, standardize = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{dvs}{A character vector of one or more Dependent Variables.}

\item{ivs}{A character vector of one or more Independent Variables.}

\item{standardize}{Logical. If TRUE, standardizes data before analysis.}
}
\value{
A list with two dataframes: 'Regression' and 'Fit'.
}
\description{
Fits a lavaan model and returns a list containing the regression
coefficients table and the global model fit indices.
}
