% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_normality_table.R
\name{make_normality_table}
\alias{make_normality_table}
\title{Make Normality Table}
\usage{
make_normality_table(data, vars = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Character vector of variable names. If NULL, all columns are used.}
}
\value{
A data frame with W/D statistics and p-values for both tests.
}
\description{
Computes Shapiro-Wilk and Kolmogorov-Smirnov normality tests for specified variables.
}
