% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reverse_score.R
\name{make_reverse_score}
\alias{make_reverse_score}
\title{Create Reverse Scores as New Columns}
\usage{
make_reverse_score(data, vars, min_val, max_val, suffix = "_rev")
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of column names to reverse.}

\item{min_val}{The minimum possible score of the scale.}

\item{max_val}{The maximum possible score of the scale.}

\item{suffix}{A character string to append to the new column names. Defaults to "_rev".}
}
\value{
A data frame containing the original data plus the new reversed columns.
}
\description{
Create Reverse Scores as New Columns
}
