% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scale_total.R
\name{make_scale_total}
\alias{make_scale_total}
\title{Compute Scale Total Score}
\usage{
make_scale_total(data, vars, new_var)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of column names to be summed.}

\item{new_var}{A single character string specifying the name of the
new total score column.}
}
\value{
The input data frame with one additional numeric column.
}
\description{
Computes the total score for a psychometric scale by summing specified
numeric variables. The resulting total score is appended as a new column
to the input data frame. Missing values are handled using `na.rm = TRUE`.
}
